requirements.txt
psycopg2-binary
pyyaml
uuid
requests